fix: 拆分用户验证函数
